e2e: Refactor `BeforeAll` section
Move the device discovery and `BeforeAll` checks in the `BeforeSuite`
section of the test suite.
