+++ test/conformance/test_suite_test.go
@@ -17,7 +17,6 @@ import (
 	// Test files in this package must not end with `_test.go` suffix, as they are imported as go package
 	_ "github.com/k8snetworkplumbingwg/sriov-network-operator/test/conformance/tests"
 
-	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/clean"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/k8sreporter"
 )
 
@@ -59,12 +58,3 @@ func TestTest(t *testing.T) {
 	RunSpecs(t, "SRIOV Operator conformance tests")
 }
 
-var _ = BeforeSuite(func() {
-	err := clean.All()
-	Expect(err).NotTo(HaveOccurred())
-})
-
-var _ = AfterSuite(func() {
-	err := clean.All()
-	Expect(err).NotTo(HaveOccurred())
-})
