+++ test/conformance/tests/fixtures.go
@@ -1,0 +1,49 @@
+package tests
+
+import (
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/clean"
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/cluster"
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/discovery"
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/namespaces"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+var sriovInfos *cluster.EnabledNodes
+
+
+var _ = BeforeSuite(func() {
+	err := clean.All()
+	Expect(err).NotTo(HaveOccurred())
+
+	isSingleNode, err := cluster.IsSingleNode(clients)
+	Expect(err).ToNot(HaveOccurred())
+	if isSingleNode {
+		disableDrainState, err := cluster.GetNodeDrainState(clients, operatorNamespace)
+		Expect(err).ToNot(HaveOccurred())
+		if discovery.Enabled() {
+			if !disableDrainState {
+				Fail("SriovOperatorConfig DisableDrain property must be enabled in a single node environment")
+			}
+		}
+		snoTimeoutMultiplier = 1
+		if !disableDrainState {
+			err = cluster.SetDisableNodeDrainState(clients, operatorNamespace, true)
+			Expect(err).ToNot(HaveOccurred())
+			clean.RestoreNodeDrainState = true
+		}
+	}
+
+	err = namespaces.Create(namespaces.Test, clients)
+	Expect(err).ToNot(HaveOccurred())
+	err = namespaces.Clean(operatorNamespace, namespaces.Test, clients, discovery.Enabled())
+	Expect(err).ToNot(HaveOccurred())
+	WaitForSRIOVStable()
+	sriovInfos, err = cluster.DiscoverSriov(clients, operatorNamespace)
+	Expect(err).ToNot(HaveOccurred())
+})
+
+var _ = AfterSuite(func() {
+	err := clean.All()
+	Expect(err).NotTo(HaveOccurred())
+})
