+++ test/conformance/tests/test_policy_configuration.go
@@ -15,7 +15,6 @@ import (
 	netattdefv1 "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1"
 
 	sriovv1 "github.com/k8snetworkplumbingwg/sriov-network-operator/api/v1"
-	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/clean"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/cluster"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/discovery"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/namespaces"
@@ -24,36 +23,6 @@ import (
 )
 
 var _ = Describe("[sriov] operator", Ordered, func() {
-	var sriovInfos *cluster.EnabledNodes
-
-	BeforeAll(func() {
-		isSingleNode, err := cluster.IsSingleNode(clients)
-		Expect(err).ToNot(HaveOccurred())
-		if isSingleNode {
-			disableDrainState, err := cluster.GetNodeDrainState(clients, operatorNamespace)
-			Expect(err).ToNot(HaveOccurred())
-			if discovery.Enabled() {
-				if !disableDrainState {
-					Skip("SriovOperatorConfig DisableDrain property must be enabled in a single node environment")
-				}
-			}
-			snoTimeoutMultiplier = 1
-			if !disableDrainState {
-				err = cluster.SetDisableNodeDrainState(clients, operatorNamespace, true)
-				Expect(err).ToNot(HaveOccurred())
-				clean.RestoreNodeDrainState = true
-			}
-		}
-
-		err = namespaces.Create(namespaces.Test, clients)
-		Expect(err).ToNot(HaveOccurred())
-		err = namespaces.Clean(operatorNamespace, namespaces.Test, clients, discovery.Enabled())
-		Expect(err).ToNot(HaveOccurred())
-		WaitForSRIOVStable()
-		sriovInfos, err = cluster.DiscoverSriov(clients, operatorNamespace)
-		Expect(err).ToNot(HaveOccurred())
-	})
-
 	Describe("Custom SriovNetworkNodePolicy", func() {
 		BeforeEach(func() {
 			err := namespaces.Clean(operatorNamespace, namespaces.Test, clients, discovery.Enabled())
